+++ spec/fixtures/class_publisher.rb
@@ -2,8 +2,18 @@
 
 require_relative '../../lib/observers'
 
+class InverseSubscriber
+  class << self
+    def action(event: nil) # rubocop:disable Lint/UnusedMethodArgument
+      nil # Stubbed via test but here to show intent.
+    end
+  end
+end
+
 class ClassPublisher
   include Observers
+
+  observers << InverseSubscriber
 
   class << self
     def trigger_action(action:)